requirements.txt
# ACRB Requirements - December 2025 Latest Models
# Install: pip install -r requirements.txt

# Core Deep Learning
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.47.0
accelerate>=1.2.0

# Diffusion Models (Dec 2025 SOTA)
diffusers>=0.32.0
safetensors>=0.4.5
peft>=0.14.0  # For LoRA support

# Vision-Language Models
# Qwen2.5-VL for cue retention scoring (LOCAL, FREE)
qwen-vl-utils>=0.0.8
# OpenAI for API-based VLM (optional)
openai>=1.58.0

# CLIP for refusal detection
git+https://github.com/openai/CLIP.git
open-clip-torch>=2.28.0

# Image Processing
pillow>=11.0.0
opencv-python>=4.10.0

# Data Processing
numpy>=2.0.0
pandas>=2.2.0
scipy>=1.14.0

# Utilities
tqdm>=4.67.0
python-dotenv>=1.0.0
requests>=2.32.0
huggingface-hub>=0.27.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0

# Optional: API Clients (for paid models)
# google-generativeai>=0.8.0
# anthropic>=0.40.0
